requirements.txt
fbchat
python-dotenv
schedule
sqlalchemy

